Sensores_Iot/Iot-A6/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
aiocoap==0.4.4
cryptography==41.0.7
requests==2.31.0
python-multipart==0.0.6
PyJWT==2.8.0
python-dotenv==1.0.0
websockets